changed schema for Card and Selection
Removed unnecessary column from Selection - the card order is unnecessary, so just save the chosen card and the other two options.

Card - there's no need to store the entire hearthpwn url as a string, as it's identical other than the ID in the url. Just store that as an integer.